Bugs with "Error" Result and Parentheses Fixed
Now, when a button is pressed and has an "Error" result on the screen, the Error Text will disappear automatically. Furthermore, when an Open Parenthesis is the last character, only some operations can be used immediately afterwards. And finally, if there are some parentheses that were not closed when the "=" button is pressed, they will be automatically closed.