Local Storage and Bug fixed in SimpleCalculator
Bug showing "NaN" and "Infinity" fixed (Now it just shows an "Error").
DRY in the Calculator's constructor with the "clearAll" method.
The use of a "class" in the buttons container changed to an "id" for performance.
Now the calculator can save the user's preferred theme in local storage and automatically checks their system preference.
More explanatory comments added.